// src/schemas.ts
export interface StoredCacheEntry {

  readonly status: number;

  readonly headers: Map<string, string>;

  readonly body: Uint8Array;

  readonly storedAt: Date;

  readonly maxAge: number;

  readonly swr: number;

  readonly tags: string[];

  readonly path: string;

  readonly vary: string[];

  readonly varyValues: Map<string, string>;
}

export const StoredCacheEntry = /*#__PURE__*/ Object.freeze(/*#__PURE__*/ Object.assign(
  // Factory function
  (data: StoredCacheEntry): StoredCacheEntry & BebopRecord => {
    return Object.freeze({
      ...data,
      encode(): Uint8Array {
        return StoredCacheEntry.encode(this);
      }
    });
  },
  // Static methods
  {
    encode(record: StoredCacheEntry): Uint8Array {
      const view = BebopView.getInstance();
      view.startWriting();
      StoredCacheEntry.encodeInto(record, view);
      return view.toArray();
    },

    encodeInto(record: StoredCacheEntry, view: BebopView): void {
      view.writeUint16(record.status);
      view.writeUint32(record.headers.size);
      for (const [k0, v0] of record.headers) {
        view.writeString(k0);
        view.writeString(v0);
      }
      view.writeBytes(record.body);
      view.writeDate(record.storedAt);
      view.writeUint32(record.maxAge);
      view.writeUint32(record.swr);
      {
        const length0 = record.tags.length;
        view.writeUint32(length0);
        for (let i0 = 0; i0 < length0; i0++) {
          view.writeString(record.tags[i0]);
        }
      }
      view.writeString(record.path);
      {
        const length0 = record.vary.length;
        view.writeUint32(length0);
        for (let i0 = 0; i0 < length0; i0++) {
          view.writeString(record.vary[i0]);
        }
      }
      view.writeUint32(record.varyValues.size);
      for (const [k0, v0] of record.varyValues) {
        view.writeString(k0);
        view.writeString(v0);
      }
    },

    decode(buffer: Uint8Array): StoredCacheEntry & BebopRecord {
      const view = BebopView.getInstance();
      view.startReading(buffer);
      const decoded = StoredCacheEntry.readFrom(view);
      return StoredCacheEntry(decoded);
    },

    readFrom(view: BebopView): StoredCacheEntry {
      let field0: number;
      field0 = view.readUint16();
      let field1: Map<string, string>;
      {
        const length0 = view.readUint32();
        field1 = new Map();
        for (let i0 = 0; i0 < length0; i0++) {
          let k0: string;
          let v0: string;
          k0 = view.readString();
          v0 = view.readString();
          field1.set(k0, v0);
        }
      }
      let field2: Uint8Array;
      field2 = view.readBytes();
      let field3: Date;
      field3 = view.readDate();
      let field4: number;
      field4 = view.readUint32();
      let field5: number;
      field5 = view.readUint32();
      let field6: string[];
      {
        const length0 = view.readUint32();
        field6 = [];
        for (let i0 = 0; i0 < length0; i0++) {
          let x0: string;
          x0 = view.readString();
          field6[i0] = x0;
        }
      }
      let field7: string;
      field7 = view.readString();
      let field8: string[];
      {
        const length0 = view.readUint32();
        field8 = [];
        for (let i0 = 0; i0 < length0; i0++) {
          let x0: string;
          x0 = view.readString();
          field8[i0] = x0;
        }
      }
      let field9: Map<string, string>;
      {
        const length0 = view.readUint32();
        field9 = new Map();
        for (let i0 = 0; i0 < length0; i0++) {
          let k0: string;
          let v0: string;
          k0 = view.readString();
          v0 = view.readString();
          field9.set(k0, v0);
        }
      }
      return {
        status: field0,
        headers: field1,
        body: field2,
        storedAt: field3,
        maxAge: field4,
        swr: field5,
        tags: field6,
        path: field7,
        vary: field8,
        varyValues: field9,
      };
    },
  }
));